adjust format function signature to accept seconds as long instead of int to avoid any potential precision loss.
make slight correction to format method and setw(2) for seconds as well so we don't have single digit seconds

+++ include/format.h
@@ -7,6 +7,6 @@
 #include <cmath>
 
 namespace Format {
-    std::string ElapsedTime(int times);  
+    std::string ElapsedTime(long times);  
 };                                    
 #endif
